% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banxico_ts.R
\name{banxico_ts}
\alias{banxico_ts}
\title{Banxico API}
\usage{
banxico_ts(series_id_banxico, initial_date, final_date)
}
\arguments{
\item{final_date}{}
}
\value{
Return time series from Economic Information System of Banco de Mexico.
}
\description{
This function is useful to import quickly data sets from Banco de Mexico.
You just need the to type the initial date, final date, and series id. All the parameters
with quotation marks.
}
\examples{
banxico_ts("SF1","2000-01-01","2020-01-01")
}
