% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_boots.R
\name{block_boots}
\alias{block_boots}
\title{Block Bootstrapping}
\usage{
block_boots(k, B, data_bb)
}
\arguments{
\item{k}{Number of blocks}

\item{B}{Bootstraps}

\item{data_bb}{Data in data frame format. Only two variables.}
}
\value{

}
\description{
This function is to perform block bootstrapping which is more convenient
for time series. This function only allows correlations by the moment, but it will be
improved hopefully during Winter Term.
}
