% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boots.R
\name{boots}
\alias{boots}
\title{Bootstrapping}
\usage{
boots(B, data)
}
\arguments{
\item{B}{Bootstraps}

\item{data}{Data in data frame.}
}
\value{

}
\description{
This function is to perform bootstrapping and only works for correlations by
the moment, but it will be improved hopefully during Winter Term.
}
