% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahal_outlier.R
\name{mahal_outlier}
\alias{mahal_outlier}
\title{Mahalanobis Distance Outlier Detection}
\usage{
mahal_outlier(data, p_val_chi = 0.001)
}
\arguments{
\item{data}{Matrix format}

\item{p_val_chi}{Level of significance to reject Ho. It is automatically set to 0.1 percent.}
}
\value{
Outliers from your multivariate space.
}
\description{
This function is to detect outliers in the Euclidean space.
You need to convert your data in matrix format.
}
\examples{

}
