% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoneyData.R
\docType{data}
\name{MoneyData}
\alias{MoneyData}
\title{Money Demand Model for Mexico}
\format{
295 x 7 data frame

\describe{
\item{Currency in Circulation}
\item{Global Production Indicator}
\item{90 day Treasury Certificates (CETES)}
\item{Prices Index}
\item{Remittances to Mexico}
\item{Exchange Rate FIX}
}
}
\usage{
MoneyData
}
\description{
Money Demand Model for Mexico
}
\keyword{datasets}
