% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols.R
\name{ols}
\alias{ols}
\title{Ordinary Least Square Estimator}
\usage{
ols(data_ols)
}
\arguments{
\item{data_ols}{This is a function that estimates the OLS estimators and computes
other important statistics}
}
\value{

}
\description{
This function estimates the parameters by Ordinary Least Squares. This
was created manually with matrix algebra and it is useful to create other statistics inside
the function.
}
