% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_residuals.R
\name{ols_residuals}
\alias{ols_residuals}
\title{Jarque Bera Test for Residuals}
\usage{
ols_residuals(model, cl)
}
\arguments{
\item{model}{results from ols or ols_se_white  function}

\item{cl}{}
}
\value{

}
\description{
This function provides useful graphs of the residuals, as well as the Jarque Bera Test
calculated by hand.
}
