% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_se_white.R
\name{ols_se_white}
\alias{ols_se_white}
\title{OLS with White Robust Standard Errors}
\usage{
ols_se_white(data_ols)
}
\arguments{
\item{data_ols}{Data in matrix format}
}
\value{

}
\description{
This function is the same as OLS, but I calculated the White standard errors
}
