% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_ols.R
\name{sum_ols}
\alias{sum_ols}
\title{OLS Main Results}
\usage{
sum_ols(data_sum)
}
\arguments{
\item{data_sum}{Results from ols function or ols_se_white}
}
\value{

}
\description{
This function returns useful statistics from the regression analysis
calculated by hand.
}
