% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif_mc.R
\name{vif_mc}
\alias{vif_mc}
\title{Variance Inflation Factor}
\usage{
vif_mc(data_vif)
}
\arguments{
\item{data_vif}{Data in matrix format. Only include the regressors. Omit the dependent variable.}
}
\value{

}
\description{
This function is to detect Multicollinearity among the explanatory variables.
}
